refactor: move response function
